requirements.txt
flask==2.0.1
pandas==1.5.3
numpy==1.21.0
python-dotenv==0.19.0
kaggle==1.5.12
scikit-learn==0.24.2
tqdm==4.62.0
scipy==1.7.0
matplotlib==3.4.2
seaborn==0.11.1
jupyter==1.0.0